backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
requests
python-dotenv
motor
pydantic
